backend/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0